+++ src/main/java/br/com/integratefarma/fornecedor/entity/FornecedorEntity.java
@@ -22,7 +22,7 @@ public class FornecedorEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "nome")
     private String nome;
@@ -46,7 +46,7 @@ public class FornecedorEntity {
     private String endereco;
 
     @Column(name = "numero")
-    private Long numero;
+    private Integer numero;
 
     @Column(name = "complemento")
     private String complemento;
