+++ src/main/java/br/com/integratefarma/funcionario/dto/FuncionarioDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class FuncionarioDTO {
-    private Long id;
+    private Integer id;
     private String nome;
     private String rg;
     private String cpf;
@@ -21,7 +21,7 @@ public class FuncionarioDTO {
     private String celular;
     private String cep;
     private String endereco;
-    private Long numero;
+    private Integer numero;
     private String complemento;
     private String bairro;
     private String cidade;
